Support TLS
It is considered a good practice to encrypt traffic. In our case, when using Home Assistant we do not need to because we expect the network to be private but bad actors can occur. Some people may want to actually to have it encrypted.
Dealing with certificates and PKI has never been fun for most of us but it has improved much in the last 5 or so years.
Since Mosquitto supports this and some people have it setup, we keep the support.

Unfortunately, there is no easy way to get this going for UNO WiFi Rev2 and UNO R4 WiFi boards.